Move StrvGetType into internal namespace
