+++ src/Libs/GObject-2.0/Internal/Functions.cs
@@ -1,6 +1,6 @@
 ﻿using System.Runtime.InteropServices;
 
-namespace GObject;
+namespace GObject.Internal;
 
 public partial class Functions
 {
